Adjust fixer role to the RHOSO environment
In the RHOSO env, where control plane is podified, things like e.g.
names of the systemd services or location of the config files are
different from what we have in the environments based on the Tripleo.
This patch adjusts that fixer role to be working fine in both
environment types.