Add helper script to handle common functions for the workshop versions
Workshop script for the podified (RHOSO) environments will be
slightly different than the existing one which is designed to be used
on the Tripleo based deployments, but some functions and variables are
the same for both scripts, so we can extract those common parts to
separate file and reuse them in both scripts.

This patch introduces that common library which is now used in the
existing osp-workshop-2022.sh script and will be also used later in the
new version of the script.