Use 'debug' variable to check where ansible output should be send
Previously osp-workshop script was only checking is $ansible_params
variable is equal to '-vv' and based on that decided if ansible-playbook
output should be displayed or send to the /dev/null.
Now this variable can contain more parameters so this patch added
'debug' variable which is used to check where to send ansible-playbook
output.