Adjust 'scenario 1' role for the podified deployment
This patch adds variables like 'oc_cmd' and 'oc_namespace' to the
ansible role for the scenario 1. Those roles are used to run all
openstack commands as in the podified deployment those commands should
be run using openstackclient POD.
It also delegates all tasks from that role to be run on localhost in
case of the podified deployment.