Remove ansible ssh key from the inventory file
In case of the podified environment, if inventory file don't exists,
script will get it from the Openshift. But this file is actually
prepared to be used from the ansible POD and have ssh key defined as in
the /runner/... location.
When osp-workshop-2024 script is run, we should use ssh key provided
with the '-p' parameter to the script so the default key from the
inventory file is now deleted.