Rename oc_cmd variable to oc_osp_cmd in scenario 1 and 2 roles
As in scenario 4 there is other command for the 'oc exec' needed to run
command in other pod than openstackclient, oc_cmd var is used for that
and var which holds command to run openstack client command is now named
'oc_osp_cmd'.
This patch renames existing 'oc_cmd' var in scenarios 1 and 2 roles to
be consistent across all roles.