Fix osp-workshop-2022.sh script
After changes which were done to adopt workshop script to the RHOSO
environments and after moving parts of the script to the common module,
there were some missing updates in the original script for the tripleo
environment. This patch updated tripleo related version of the script to
be similar as the 2024 version.