Run tasks which are supposed to be run on compute nodes only there
Previously tasks which were supposed to be run on the compute nodes
where run on nodes from the 'compute_group_name' inventory group if the
installer was set to 'tripleo' but the same task was going to be run on
all the nodes from the inventory in the podified environment.

Now this is changed and those tasks will be run only on the nodes from
the 'compute_group_name' group no matter of the installer used.