Change name of the parameter to provide inventory secret name
It was previously '-f' option used to provide name of the secret where
the ansible inventory is stored in Openshift. But as the same '-f' is
used in the 2022 version of script to provide inventory file name in the
tripleo based deployments, this patch changes it to be '-s' option in
2024 name to provide name of the secret.

Additionally this patch cleans leftover 'u' argument from the list of accepted
arguments in the 2024 script as this is related to undercloud so useful
only in the 2022 script version.