Add scenario13 with break/fix of the FIP port forwarding
This patch adds new role for the scenario13 which is focused on the
floating ip port forwarding.
Scenario first creates "vlan" tenant network and attaches 2 VMs to it.
One of the VMs have Floating IP associated (this is distributed) and
other have FIP Port Forwarding (this traffic should be centralized).
Such configuration is not supported by Neutron thus FIP PF's connection
don't work.

As fix for that issue, there is new "geneve" network created and VMs
are recreated and connected to that new network.

Closes: #OSPRH-6202