Add script to prepare simple inventory file based for the training env
This script is taking EDPM inventory from the Openshift and creates very
simple inventory file with list of hosts, ansible_user and ansible_host
set for each host.
The reason why inventory from Openshift can't be taken "as is" is the
fact that in the lab environment an "external_ip" needs to be used to
access EDPM nodes from the workstation node and this needs to be changed
comparing to the EDPfm inventory file.