Add way to configure ProxyJump host for the ssh connections
In the training lab environment edpm-compute nodes can be accessed
with SSH through the "utility" server, not directly. So this patch adds
possibility to configure such ProxyJump host for each edpm node from the
ansible inventory file. This is going to be configured when jump host is
given with "-j" parameter to the osp-workshop-2024.sh script