[Scenario 2] Fix creation of the "broken" image
Previously broken (empty) image created in the scenario 2 was created as
simple empty file with size of 0 bytes and then uploaded to Glance.
It seems that in RHOSO 18 such image can't be used to spawn VM as it
complains that "image is not valid" and VM is set to ERROR state.

To avoid that, image will now be created as empty qcow2 image and
upload to Glance.