Rename scenario roles and task files according to the lab's convention
This patch renames scenarios used currently in the lab:

* scenario1 to scenariobfx019
* scenario2 to scenariobfx020
* scenario4 to scenariobfx021
* scenario5 to scenariobfx022
* scenario6 to scenariobfx023
* scenario9 to scenariobfx026
* scenario13 to scenariobfx027

It also renames task files and tags to match the ones used in the
training lab, so there are tasks (and tags) like:

* "start" instead of "break",
* "finish" instead of "cleanup"
* "grade" and "fix" which are not changed.

Finally this patch also changes things like e.g. internal subnet ranges
used by subnets in each of those scenarios.

This patch updates only scenarios mentioned above as those are prepared
and will be currently used for the training lab. Other scenarios needs
to be adjusted in future if that will be necessary.