Add configuation of the nicMappings in the ovnController POD
To be able to use ovnController PODs running on the OpenShift nodes as
gateway chassis there need to be `nicMappings` configured in the
OpenStackControlPlane CR. This is missing (for now) in the lab
deployment template so the `lab` script will add it as part of the
preparation of the lab environment.
Currently there is only one scenarion (bfx027) which really needs it,
all other scenarios are relying on the distributed Floating IPs and
connectivity directly to/from compute nodes but configuring those
nicMappings will not break any of those scenarios and helps with the
bfx027.

There is no problem with doing this change on start of every scenario as
OpenShift is smart enough and will not actually do any changes if the
nicMapping was already set to the same value.

Related: OSPRH-14076