Set jump host in lab script and make sure no old fingerprint exists
Previously SSH connection to the compute nodes through the lab@utility
server as jump host was only configured by the osp-workshop helper
script but as this is really needed on the lab environment for all of
the scenarios run with RHOSO 18 this patch adds configuration of the
jump host for the compute nodes also in the the lab script.

Additionally this patch removes existing ~/.ssh/known_hosts file to make
sure that there is not issue with host keys verification after e.g. lab
stop and then started again.