Unify inventory filename and DATADIR in lab and osp-workshop scripts
This patch sets names of the inventory file name and DATADIR used by the
osp-workshop-2024.sh and lab scripts to be the same. It is useful as
osp-workshop-2024.sh script may be useful to create ansible inventory
file which then will be used by the lab script.

Additionally this patch moves configure_ssh_jump_host function's
implementation in the lab script to be done before it is actually used
in that script.